+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@ Changes:
   * Fixed an issue that allowed login to time out during bulk save
   * Fixed remember-me token not being checked during REST requests (during saves, searches, etc.),
     causing the requests to fail authentication if the session had timed out
+  * Added an editor for printer labels
 
 # 1.16.0